Ouroboros.Network.Testing.Data.Script

Test scripts

newtype Script a

Constructors

Script (NonEmpty a)

data NonEmpty a

Non-empty (and non-strict) list type.

Since: base-4.9.0.0

Constructors

a :| [a] infixr 5

scriptHead ∷ Script a → a

singletonScript ∷ a → Script a

initScript ∷ MonadSTM m ⇒ Script a → STM m (TVar m (Script a))

stepScript ∷ MonadSTM m ⇒ TVar m (Script a) → m a

stepScriptSTM ∷ MonadSTM m ⇒ TVar m (Script a) → STM m a

initScript' ∷ MonadSTM m ⇒ Script a → m (TVar m (Script a))

stepScript' ∷ MonadSTM m ⇒ TVar m (Script a) → m a

stepScriptSTM' ∷ MonadSTM m ⇒ TVar m (Script a) → STM m a

arbitraryScriptOf ∷ Int → Gen a → Gen (Script a)

prop_shrink_Script ∷ Fixed (Script Int) → Property

Timed scripts

data ScriptDelay

Constructors

NoDelay

ShortDelay

LongDelay

type TimedScript a = Script (a, ScriptDelay)

playTimedScript ∷ (MonadAsync m, MonadTimer m) ⇒ Tracer m a → TimedScript a → m (TVar m a)

Pick scripts

type PickScript peeraddr = Script (PickMembers peeraddr)

A pick script is used to interpret the policyPickKnownPeersForGossip and the policyPickColdPeersToForget. It selects elements from the given choices by their index (modulo the number of choices). This representation was chosen because it allows easy shrinking.

data PickMembers peeraddr

Constructors

PickFirst

PickAll

PickSome (Set peeraddr)

arbitraryPickScript ∷ Gen (Set peeraddr) → Gen (PickScript peeraddr)

interpretPickScript ∷ (MonadSTM m, Ord peeraddr) ⇒ TVar m (PickScript peeraddr) → Set peeraddr → Int → STM m (Set peeraddr)