Ouroboros.Network.Testing.Utils

Arbitrary Delays

newtype Delay

Constructors

Delay

Fields

getDelay :: DiffTime

genDelayWithPrecision :: Integer -> Gen DiffTime

newtype SmallDelay

Constructors

SmallDelay

Fields

getSmallDelay :: DiffTime

QuickCheck Utils

arbitrarySubset :: Ord a => Set a -> Gen (Set a)

Pick a subset of a set, using a 50:50 chance for each set element.

shrinkVector :: (a -> [a]) -> [a] -> [[a]]

Like shrinkList but only shrink the elems, don't drop elements.

Useful when you want a custom strategy for dropping elements.

prop_shrink_valid :: (Arbitrary a, Show a) => (a -> Bool) -> Fixed a -> Property

Check that each shrink satisfies some invariant or validity condition.

prop_shrink_nonequal :: (Arbitrary a, Eq a) => Fixed a -> Property

The shrink function needs to give a valid value that is smaller than the original, otherwise the shrinking procedure is not well-founded and can cycle.

This property does not check size, as that would need significant extra infrastructure to define an appropriate measure. Instead this property simply checks each shrink is not the same as the original. This catches simple 1-cycles, but not bigger cycles. These are fortunately the most common case, so it is still a useful property in practice.

Tracing Utils

data WithName name event

Constructors

WithName

Fields

wnName :: name

wnEvent :: event

data WithTime event

Constructors

WithTime

Fields

wtTime :: Time

wtEvent :: event

tracerWithName :: name -> Tracer m (WithName name a) -> Tracer m a

tracerWithTime :: MonadMonotonicTime m => Tracer m (WithTime a) -> Tracer m a

tracerWithTimeName :: MonadMonotonicTime m => name -> Tracer m (WithTime (WithName name a)) -> Tracer m a

swapTimeWithName :: WithTime (WithName name b) -> WithName name (WithTime b)

swapNameWithTime :: WithName name (WithTime b) -> WithTime (WithName name b)

splitWithNameTrace :: Ord name => Trace r (WithName name b) -> Trace r [WithName name b]

Split Trace events into separate traces indexed by a given name.

Tracers

debugTracer :: (Show a, Applicative m) => Tracer m a

sayTracer :: (Show a, MonadSay m) => Tracer m a

Tasty Utils

nightlyTest :: TestTree -> TestTree

ignoreTest :: TestTree -> TestTree

Prevents the tests from running and reports them as succeeding.

This may be be desireable as an alternative to commenting out the tests. This way, they are still typechecked (preventing bitrot), and the test report lists them, which serves as a reminder that there are ignored tests.

Note that any setup/teardown actions executed by withResource are still executed. You can bypass this manually as in the following example:

askOption $ \(MyFlag b) -> if b
                           then withResource mytest
                           else ignoreTest . mytest $ return junkvalue

Auxiliary functions

renderRanges :: Int -> Int -> String

Use in tabulate to help summarise data into buckets.