Ouroboros.Network.Testing.Utils

Documentation

newtype Delay

Constructors

Delay

Fields

getDelay ∷ DiffTime

genDelayWithPrecision ∷ Integer → Gen DiffTime

newtype SmallDelay

Constructors

SmallDelay

Fields

getSmallDelay ∷ DiffTime

arbitrarySubset ∷ Ord a ⇒ Set a → Gen (Set a)

Pick a subset of a set, using a 50:50 chance for each set element.

shrinkVector ∷ (a → [a]) → [a] → [[a]]

Like shrinkList but only shrink the elems, don't drop elements.

Useful when you want a custom strategy for dropping elements.

prop_shrink_valid ∷ (Arbitrary a, Show a) ⇒ (a → Bool) → Fixed a → Property

Check that each shrink satisfies some invariant or validity condition.

prop_shrink_nonequal ∷ (Arbitrary a, Eq a) ⇒ Fixed a → Property

The shrink function needs to give a valid value that is smaller than the original, otherwise the shrinking procedure is not well-founded and can cycle.

This property does not check size, as that would need significant extra infrastructure to define an appropriate measure. Instead this property simply checks each shrink is not the same as the original. This catches simple 1-cycles, but not bigger cycles. These are fortunately the most common case, so it is still a useful property in practice.

renderRanges ∷ Int → Int → String

Use in tabulate to help summarise data into buckets.

debugTracer ∷ (Show a, Applicative m) ⇒ Tracer m a

sayTracer ∷ (Show a, MonadSay m) ⇒ Tracer m a